% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gbsamplerC}
\alias{gbsamplerC}
\title{A Gibbs sampler using Rcpp}
\usage{
gbsamplerC(m, a, b, n)
}
\arguments{
\item{m}{numbers of sample want to generate}

\item{a}{the a parameter in Beta(x+a,n-x+b)}

\item{b}{the b parameter in Beta(x+a,n-x+b)}

\item{n}{the n parameter in Binomial(n,y)}
}
\value{
a matrix including two random sample with dimension of (\code{m},2)
}
\description{
A specific Gibbs sampler for 2 dimension (x,y) with bivariate density using Rcpp
}
\examples{
\dontrun{
  dir_cpp <- '/data/cenmin/statistical computing/Rcpp/'
  sourceCpp(paste0(dir_cpp,"GBSampler.cpp"))
  set.seed(12)
  m = 10000
  a = 2
  b = 3
  n = 5
  t1 <- proc.time()
  xC <- gbsamplerC(m,a,b,n)
  t2 <- proc.time()
  t <- t2-t1
  paste0(t[3][[1]],'s') 
} 
}
