% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digital_pathology.R
\name{matchFeaLab}
\alias{matchFeaLab}
\title{Match Feature with label in MIL situation in digital_pathology}
\usage{
matchFeaLab(
  feature_path,
  Label,
  cancer_types,
  save_path,
  id_len = 12,
  ifclass = TRUE,
  threshold = 0.4
)
}
\arguments{
\item{feature_path}{the path of file which contains feature csv files.}

\item{Label}{a label data frame with column names 'id','label','CancerType'.}

\item{cancer_types}{a vector contains all cancer types you want to focus.}

\item{save_path}{the file path you want to save the matched 'fealabel' data frame.}

\item{id_len}{one-dimension numeric.the length of the patient id in feature file name. The default value is 12 for TCGA patients.}

\item{ifclass}{a bool value for whether attach a 0-1 label to feature,TRUE for yes,FALSE for no. The default value is TRUE.}

\item{threshold}{one-dimension numeric.the threshold for 0 or 1, when is.class==TRUE. The default value is 0.4 for MSI}
}
\value{
a list of patient id after match
}
\description{
It can directly match features with label in MIL situation in digital_pathology save the feature_lab files in "save file" for each patient.For each patient has one feature file of ten thousands tile samples and one label. The id of patient is present in file name.
}
\examples{
\dontrun{
    feature_path ="/data/cenmin/TCGA_CRC_Tumorfeaposinocsv"
    label_path = "/data/cenmin/statR/MSI.csv"
    Label = read.csv(label_path,sep="\t")
    save_path = "/data/cenmin/statR/TCGA"
    cancer_types = c("COAD","READ")
    colnames(Label) <- c('id','label','CancerType')
    save_id = matchFeaLab(feature_path,Label,cancer_types,save_path)
}
}
