\name{StatComp21097-package}
\alias{StatComp21097-package}
\docType{package}
\title{
  This package provides several functions for statsical computing and digital pathology
}
\description{
  There are some R functions for efficient digital pathology and Cpp functions of gibbs sampler.
}
\author{
Min Cen
Maintainer: Min Cen <cenmin0127@mail.utsc.edu.com>
}
\keyword{ package }
